+++ saleor/checkout/tests/fixtures/checkout.py
@@ -172,21 +172,6 @@ def checkout_ready_to_complete(
 
 
 @pytest.fixture
-def checkout_with_digital_item(checkout, digital_content, address):
-    """Create a checkout with a digital line."""
-    variant = digital_content.product_variant
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    add_variant_to_checkout(checkout_info, variant, 1)
-    checkout.discount_amount = Decimal(0)
-    checkout.billing_address = address
-    checkout.email = "customer@example.com"
-    checkout.save()
-    return checkout
-
-
-@pytest.fixture
 def checkout_with_shipping_required(checkout_with_item, product):
     checkout = checkout_with_item
     variant = product.variants.get()