+++ saleor/graphql/order/tests/mutations/test_order_update_shipping.py
@@ -924,9 +924,9 @@ def test_editable_order_update_shipping_triggers_proper_updated_webhook_for_null
     setup_order_webhooks,
     staff_api_client,
     permission_group_manage_orders,
-    # Only digital-lines order is affected, because orders with physical items
+    # Only orders without shipping are affected, because orders with physical items
     # must have shopping method. In such scenario validation will raise without webhook
-    order_with_digital_line,
+    order_without_shipping_required,
     settings,
 ):
     # given
@@ -938,7 +938,7 @@ def test_editable_order_update_shipping_triggers_proper_updated_webhook_for_null
     ) = setup_order_webhooks(WebhookEventAsyncType.ORDER_UPDATED)
 
     permission_group_manage_orders.user_set.add(staff_api_client.user)
-    order = order_with_digital_line
+    order = order_without_shipping_required
     order.base_shipping_price = zero_money(order.currency)
     order.status = OrderStatus.UNCONFIRMED
     order.save()